rename glasses component
