+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
 import Navbar from './Navbar';
-import Sunglasses from './Sunglasses';
+import Glasses from './Glasses';
 import Video from './Video';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
@@ -17,7 +17,7 @@ function App() {
     <div className="container-fluid vh-100 d-flex flex-column ">
       <Navbar />
       <Video height={500} width={500} onPoseChange={handlePoseChange}>
-        <Sunglasses height={500} width={500} pose={pose} />
+        <Glasses height={500} width={500} pose={pose} />
       </Video>
     </div>
   );
