+++ src/Glasses/index.js
@@ -1,0 +1,3 @@
+import Glasses from './Glasses';
+
+export default Glasses;
