--- src/Sunglasses/index.js
@@ -1,3 +1,0 @@
-import Sunglasses from './Sunglasses';
-
-export default Sunglasses;
